refactor: default bf16 storage, int8 is future extension
- MambaStatePoolConfig.compress_dtype: int8 → bfloat16 (default)
- MambaStateConfig.enable_int8_compress: True → False (default off)
- checkpoint_pool: store_from_active/load_to_active branch on compress_dtype
  (bf16 = direct copy, int8 = quantize/dequantize)
- Remove FLEXKV_MAMBA_ENABLE_INT8 env var from GLOBAL_CONFIG_FROM_ENV
- Update slot_bytes calc: use dtype_bytes (bf16=2) instead of int8=1+scale
- int8 quantize/dequantize code preserved for future use

+++ flexkv/mamba_state/config.py
@@ -32,7 +32,9 @@ class MambaStatePoolConfig:
     evict_ratio: float = 0.1
 
     # --- compression ---
-    compress_dtype: torch.dtype = torch.int8
+    # bf16: lossless, aligned with sglang HiCache host pool (default).
+    # int8: 2x capacity, aligned with sglang Int8CheckpointStore (future extension).
+    compress_dtype: torch.dtype = torch.bfloat16
     scale_dtype: torch.dtype = torch.bfloat16
 
     @property
